static libraries

+++ Obsidian/src/Obsidian/ImGui/ImGuiLayer.cpp
@@ -6,9 +6,6 @@
 #include "examples/imgui_impl_opengl3.cpp"
 #include "examples/imgui_impl_glfw.cpp"
 
-#include <GLFW/glfw3.h>
-#include <glad/glad.h>
-
 namespace Obsidian {
 
 	ImGuiLayer::ImGuiLayer()
@@ -63,7 +60,7 @@ namespace Obsidian {
 		ImGuiIO& io = ImGui::GetIO();
 		Application& app = Application::Get();
 
-		io.DisplaySize = ImVec2(app.GetWindow().GetWidth(), app.GetWindow().GetHeight());
+		io.DisplaySize = ImVec2((float)app.GetWindow().GetWidth(), (float)app.GetWindow().GetHeight());
 
 		ImGui::Render();
 		ImGui_ImplOpenGL3_RenderDrawData(ImGui::GetDrawData());
